DB migration setup

+++ database/migrations/2021_06_14_114747_create_user_experiences_table.php
@@ -12,7 +12,7 @@ class CreateUserExperiencesTable extends Migration
             $table->id();
             $table->string('email');
             $table->unsignedInteger('user_id')->nullable();
-            $table->smallInteger('amount');
+            $table->integer('amount');
             $table->string('type');
             $table->timestamps();
